+++ package.json
@@ -47,7 +47,7 @@
     "tap": "^10.0.1"
   },
   "engines": {
-    "node": ">=0.10",
+    "node": ">=8",
     "npm": ">=2"
   },
   "files": [
